Essentia.jl

Documentation for Essentia.jl

Module Index

Essentia.Algorithm

Essentia.EssentiaException

Essentia.Pool

Essentia.es2julia

Essentia.jj

Essentia.julia2es_number

Essentia.rollup

Essentia.setCppOutput!

Essentia.typeInfoToStr

Detailed API

Essentia.Algorithm — Type

Functor representing an algorithm.

Instantiating

function Algorithm(name, params...; type="standard")

This function returns the algorithm name using essentia::standard or essentia::streaming depending on type. Other arguments can be passed in a key-value fashion.

Fields

name::String
type::String
algo::T
ninp::Int32
nout::Int32

Example

Essentia.Algorithm("MFCC", :dctType=>3, :logType="natural")

Running

function (
    self::Algorithm)(inputs::Pair{String, T}...
) where T

function (self::Algorithm)(inputs::T...) where T

function (
    self::Algorithm)(inputs::Tuple{Vector{Pair{String, T}}, Vector{V}}
) where T

Executes the algorithm. Note that while this function is running, the garbage-collector is suspended!

Arguments

a variadic number of pairs where:

keys must be strings with the same name as the Essentia documentation

values are C++ or Julia objects

OR

a variadic number of C++ or Julia objects in the same order as Essentia documentation

OR

[meant to be used internally] a Tuple{Vector{Pairs{String, T}}, Vector{V}} where:

pair keys are strings with the same name as Essentia documentation

pair values are C++ objects

V is a type descriptor

Returns

a Tuple{Vector{Pairs{String, T}}, Vector{V}} where:

pair keys are strings with the same name as Essentia documentation

pair values are C++ objects

V is a type descriptor

Use jj function to get a dictionary of Julia objects

Essentia.EssentiaException — Type

struct EssentiaException <: Exception
    message::String
end

An exception for something that happens in Essentia

Essentia.Pool — Type

Type definition for representing essentia::Pool using dictionaries, so that: * keys are always the following: * 'real' * 'vectorreal' * 'string * 'vectorstring' * 'matrix' * 'stereo' * 'singlereal' * 'singlevectorreal' * 'singlestring * 'singlevectorstring'

* values are `Dict{String, T}` , where `T` changes
    according to the key (see Pool C++
    [docs](https://essentia.upf.edu/doxygen/classessentia_1_1Pool.html))

Note that "stereo" corresponds to Essentia's StereSample which refer to C++ std::vector<StereoSample> and are converted into Matrix{Float32} in Julia.

Base.convert — Method

Convert essentia::Pool to our Pool type.

Used internally.

Base.convert — Method

Convert Julia's Vector{Vector{T}} to Essentia's std::vector<std::vector<T>> and vice-versa

Used internally.

Base.convert — Method

Convert Julia's Matrix{T} to Essentia's Matrix (aka TNT::Array2D) and vice-versa

Used internally.

N.B. The C++ object is not garbage-collected and some reference to x must exist so that C++ can use the output array. If x doesn't exist, the data are lost.

Base.convert — Method

Convert Julia's Tuple{T} to Essentia's Tuple2<T> and vice-versa

Used internally.

E.g. StereoSample typedef

Base.convert — Method

Convert Julia's Complex{T} to C++'s complex<T> and vice-versa

Used internally.

Base.convert — Method

Convert Julia's Matrix{T} to Essentia's vector<Tuple2<T>> and vice-versa

Used internally.

Base.convert — Method

Convert C++ std::map<std::string, T> to Dict{String, T} type.

Used internally.

Essentia.es2julia — Method

function es2julia(d::T, type_str::String) where T

Perform conversion from a C++ pointer to Julia according to the description in type_str (returned by typeInfoToStr)

The return type changes according to type_str, so this function is NOT guaranteed to be type coherent, even though it should be so.

Essentia.jj — Method

function jj(objects::Tuple{Vector{Pair}, V})::Dict where V

Takes the output of an Algorithm and converts them to Julia dictionary so that: * keys are strings with the names in Essentia documentation * values are Julia objects

If no conversion is implemented, then the Cxx object wrapping the C++ type is returned. When working with Pool, expect to receive a Cxx object and to work with icxx macro.

Essentia.julia2es_number — Method

function julia2es(d::T) where T

Perform conversion from Julia to Essentia types.

Note that in essentia, strings are threated as std::string, while icxx converts them; for this reason, you need to use this function to deal with Essentia's strings.

The return type changes according to T, so this function is type coherent

Essentia.rollup — Method

function rollup(::Type{T},
                fn::Function,
                z::AbstractArray{K},
                ws::Integer,
                hs::Integer,
                padding::String="minimum",
                padding_fill::K=0.0) where {T, K<:Number}

Execute a function over frames extracted as views of an array

Arguments

T: the type returned by fn

fn: a function accepting a frame – i.e. an n-dimensional Array – e.g. an Algorithm

z: a n-dimensional Array from which each frame is extracted; frames are extracted along rows. For instance, let A be a 5×2 array (5 samples and 2 features), and ws=3, hs=2; this function extracts frames in this way: ```julia julia> A = reshape(collect(1:10), 5, 2) 5×2 Array{Int64,2}: 1 6 2 7 3 8 4 9 5 10

  julia> selectdim(A, 1, 1:3)
  3×2 view(::Array{Int64,2}, 1:3, :) with eltype Int64:
   1  6
   2  7
   3  8

  julia> selectdim(A, 1, 3:5)
  3×2 view(::Array{Int64,2}, 3:5, :) with eltype Int64:
   3   8
   4   9
   5  10
  ```

ws: window size

hs: hop-size

padding: a string with the following possible values:

none: no padding is added and z is shortened to the biggest number multiple of ws and ≤ length(z)

minimum: z is padded both left and right with the minimum size needed to toll over all the samples

padding_fill: a value of type K with which z is padded

To get custom paddings, consider using the PaddedViews package or the FrameCutter algorithm

Returns

if T <: AbstractArray, an Array{T} with one more dimension than the output of fn.

else Vector{T}

Each row is the output of a frame.

Essentia.setCppOutput! — Method

function setCppOutput!(outputs, algo, name, type_info, i)

Given a type_info C++ structure, allocates the corresponding object and set the output of algo

Essentia.typeInfoToStr — Method

function typeInfoToStr(type_info_ptr)::String

Given a type_info * C++ structure, returns a string which describes it (as Essentia Python bindings)